requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
pymongo==4.7.2
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.1
urllib3==2.2.1
yarl==1.9.4
